refactor: consolidate SolidState functionality and remove deprecated files
